+++ sysopt/symbolic/__init__.py
@@ -5,5 +5,4 @@
 
 from sysopt.symbolic.core import *
 from sysopt.symbolic.decision_variables import *
-from sysopt.symbolic.matrix import *
 from sysopt.symbolic.graph_evaluators import *
