+++ sysopt/symbolic/core.py
@@ -1,4 +1,4 @@
-"""Functions and factories to create symbolic variables."""
+"""Functions and classes to for symbolic manipulation."""
 import weakref
 from abc import ABCMeta, abstractmethod
 from collections import defaultdict
@@ -10,9 +10,6 @@ from typing import Union, List, Callable, Tuple, Optional, Dict, NewType, Any
 
 from sysopt.helpers import flatten, slice_to_list
 from sysopt.exceptions import InvalidShape, EvaluationError
-from sysopt.symbolic.matrix import (
-    Matrix, sparse_matrix, basis_vector
-)
 
 array = np.array
 epsilon = 1e-12
@@ -24,6 +21,60 @@ SymbolicArray = NewType('SymbolicArray',
 # ------------------------------------------------------------------------------
 # ----------------------- Array Handling and Linear Algebra---------------------
 # ------------------------------------------------------------------------------
+
+
+class Matrix(np.ndarray):
+    """View of a numpy matrix for use in expression graphs."""
+
+    def __hash__(self):
+        shape_hash = hash(self.shape)
+        data_hash = hash(tuple(self.ravel()))
+        return hash((shape_hash, data_hash))
+
+    def __cmp__(self, other):
+        return self is other
+
+    def __eq__(self, other):
+        if isinstance(other, (list, tuple)):
+            return other == self.tolist()
+
+        if isinstance(other, (float, int, complex)):
+            if self.shape == (1,):
+                return self[0] == other
+            elif self.shape == (1, 1):
+                return self[0, 0] == other
+            return False
+
+        try:
+            if self.shape != other.shape:
+                return False
+        except AttributeError:
+            return False
+        if hash(self) != hash(other):
+            return False
+        result = (self - other) == 0
+        if isinstance(result, np.ndarray):
+            return result.all()
+        else:
+            return result
+
+
+def sparse_matrix(shape: Tuple[int, int]):
+    """Create an empty sparse matrix of the given dimension.
+
+    Todo:
+        replace with an actual sparse matrix representaiton!
+
+    """
+    return np.zeros(shape, dtype=float).view(Matrix)
+
+
+def basis_vector(index, dimension):
+    """Get the corresponding standard Euclidian basis vector."""
+    e_i = np.zeros(shape=(dimension,), dtype=float).view(Matrix)
+    e_i[index] = 1
+    return e_i
+
 
 class LinearMap:
     """Functional representation of a linear operator defined by a matrix."""
